requirements.txt
streamlit
numpy
pandas
matplotlib
opencv-python-headless
tensorflow-cpu==2.16.1
requests

